Rename Token to Access Token and make it stateless.
