+++ lib/girlscout.rb
@@ -5,13 +5,13 @@ require 'excon'
 require 'girlscout/version'
 require 'girlscout/resource'
 require 'girlscout/config'
-require 'girlscout/token'
 
 require 'girlscout/concerns/has_attributes'
 require 'girlscout/concerns/has_resource'
 require 'girlscout/object'
 require 'girlscout/error'
 
+require 'girlscout/access_token'
 require 'girlscout/list'
 require 'girlscout/folder'
 require 'girlscout/mailbox'
@@ -20,4 +20,3 @@ require 'girlscout/attachment'
 require 'girlscout/thread'
 require 'girlscout/customer'
 require 'girlscout/user'
-require 'girlscout/search'
